further dedupe
